†編集† ScoreをMusicSheetに、PointをScoreにした
