+++ src/main/kotlin/game/sheet/Note.kt
@@ -1,4 +1,4 @@
-package game.score
+package game.sheet
 
 import game.drawable.Keyboard
 import game.drawable.NoteBorder
