+++ src/main/kotlin/game/sheet/SheetMusic.kt
@@ -1,0 +1,6 @@
+package game.sheet
+
+data class SheetMusic (
+   val notes: NotesGetter,
+   val songInfo: SongInfo
+)
