+++ src/main/kotlin/game/sheet/SheetMusicParser.kt
@@ -1,10 +1,10 @@
-package game.score
+package game.sheet
 
-object ScoreParser {
+object SheetMusicParser {
 
    // TODO: 実際にファイルから読み込ませる
-   fun parse(): Score {
-      return Score(
+   fun parse(): SheetMusic {
+      return SheetMusic(
          NotesGetter(List(10) { Note('w', it * 1000L + 5000L) }),
          SongInfo(
             "とても良いテスト用の譜面",
