+++ src/main/kotlin/game/sheet/SongInfo.kt
@@ -1,4 +1,4 @@
-package game.score
+package game.sheet
 
 import game.drawable.SongInfoDrawer
